feat:Strings are now fully parsable and dont require "" to work  <[ci_directive]>
__DESCRIPTION:__
Made it so variable names can only be one word
Cleaned up parse command to use POI and other error checks
Reserved chars now must be escapes with \ otherwise error
Parsing finished and tested but most likely has one edge case
Can now get if a command takes arguments or not through member function

+++ include/ccli/ccli_autocomplete.inl
@@ -160,7 +160,7 @@ namespace ccli
 	{
 		acNode *ptr = m_Root;
 		const char *temp = prefix.data();
-		int prefix_end = static_cast<int>(prefix.size());
+		size_t prefix_end = prefix.size();
 
 		// Traverse tree and check if prefix exists.
 		while (ptr)
